created config.js which contains firebase config and import it in any file in need of db instance
