+++ app/html/AddPost.html
@@ -138,7 +138,7 @@
     <script>
 
     </script>
-    <script src="../../public/js/addPost.js"></script>
+    <script type="module" src="../../public/js/addPost.js"></script>
     <script src="../../public/js/editpost.js"></script>
     <script src="../../public/js/main.js"></script>
 </body>
