+++ app/html/admin.html
@@ -42,7 +42,7 @@
                             </li>
 
                             <li class="jd-link-item">
-                                <a href="">blog</a>
+                                <a href="blog.html">blog</a>
                             </li>
                             <li class="jd-link-item">
                                 <a href="contacts.html">Contacts</a>
@@ -115,7 +115,7 @@
                         </li>
 
                         <li class="jd-link-item">
-                            <a href="">blog</a>
+                            <a href="blog.html">blog</a>
                         </li>
                         <li class="jd-link-item">
                             <a href="contacts.html">Contacts</a>
@@ -163,7 +163,7 @@
 
     </section>
     
-    <script src="../../public/js/postAdmin.js"></script>
+    <script type="module" src="../../public/js/postAdmin.js"></script>
     <script src="../../public/js/main.js"></script>
 </body>
