+++ app/html/blog.html
@@ -130,7 +130,7 @@
 
 
     </section>
-    <script src="../../public/js/post.js"></script>
+    <script type="module"  src="../../public/js/post.js"></script>
     
     <script src="../../public/js/main.js"></script>
 </body>
