+++ app/html/show.html
@@ -152,7 +152,7 @@
 
 
     </section>
-    <script src="../../public/js/article.js"></script>
+    <script type="module" src="../../public/js/article.js"></script>
     <script src="../../public/js/show.js"></script>
     <script src="../../public/js/main.js"></script>
 </body>
