+++ app/html/showPostAdmin.html
@@ -68,11 +68,13 @@
                     </span></a>
                 </div>
 
-                    <a href="">
-                        <span class="delete">
-                            <ion-icon class="deleteicon" name="trash-outline"></ion-icon> delete
-                        </span>
-                    </a>
+                    <div id="delete">
+                        <a href="">
+                            <span class="delete">
+                                <ion-icon class="deleteicon" name="trash-outline"></ion-icon> delete
+                            </span>
+                        </a>
+                    </div>
                 </div>
 
                 <div class="show-card" id="show-card">
@@ -167,7 +169,7 @@
     <script>
 
     </script>
-    <script src="../../public/js/adminArticle.js"></script>
+    <script type="module" src="../../public/js/adminArticle.js"></script>
     <script src="../../public/js/show.js"></script>
     <script src="../../public/js/main.js"></script>
 </body>
