+++ public/js/adminArticle.js
@@ -1,17 +1,8 @@
-var firebaseConfig = {
-    apiKey: "AIzaSyBG5hHsXZ-4IBTRolWhznPIPVjl-7EZPh8",
-    authDomain: "mybrand-61fba.firebaseapp.com",
-    databaseURL: "https://mybrand-61fba.firebaseio.com",
-    projectId: "mybrand-61fba",
-    storageBucket: "mybrand-61fba.appspot.com",
-    messagingSenderId: "971411413605",
-    appId: "1:971411413605:web:e9ae0257c4238b9ef9448c",
-    measurementId: "G-B9F4EYSBSS"
-};
-// Initialize Firebase
-firebase.initializeApp(firebaseConfig);
-firebase.analytics();
-const db = firebase.firestore();
+/**
+ * import firebase config file
+*/
+import {db}from './config.js';
+
 
 const id = sessionStorage.getItem('data-id');
 const showCard = document.getElementById('show-card')
@@ -42,13 +33,16 @@ function viewPost(doc) {
 
 }
 
-db.collection('posts').where(firebase.firestore.FieldPath.documentId(id), '==', id).get().then((snapshot) => {
-    snapshot.docs.forEach(doc => {
-        viewPost(doc);
-        edit(doc);
-    });
+db.collection('posts').where(firebase.firestore.FieldPath.documentId(id), '==', id).onSnapshot(snap =>{
+    let changes=snap.docChanges();
+    changes.forEach(change =>{
+        if(change.type=='added'){
+            viewPost(change.doc)
+        }
+        edit(change.doc);
+        
+    })
 })
-
 
 const editblog=document.getElementById('edit');
 
@@ -58,4 +52,15 @@ function edit(doc){
         window.location.href = '../../app/html/AddPost.html', true;
         sessionStorage=doc.id;
     })
+
+}
+
+const deleteBlog=document.getElementById('delete');
+
+function deleteArticle(doc){
+    deleteBlog.addEventListener('click',(e)=>{
+        e.preventDefault();
+        sessionStorage=doc.id;
+    })
+
 }
